Fix focus with color picker
