+++ src/components/ColorPicker/ColorPicker.module.css
@@ -1,0 +1,4 @@
+.swatch:focus-visible {
+  outline: 2px solid var(--mantine-color-text) !important;
+  outline-offset: 2px !important;
+}
